requirements.txt
deepeval>=3.0.0
groq>=1.0.0
python-dotenv>=1.0.0
